% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccdc_xy.R
\name{ccdc_func}
\alias{ccdc_func}
\title{Predict surface reflectance using CCDC algorithm (Zhu & Woodcock 2014)}
\usage{
ccdc_func(
  jdoy,
  coef_intp,
  coef_slp,
  coef_cos,
  coef_sin,
  coef_cos2,
  coef_sin2,
  coef_cos3,
  coef_sin3
)
}
\arguments{
\item{jdoy}{Julian date for which to predict surface reflectance}

\item{coef_intp}{Coefficient for overall values for the ith Landsat Band}

\item{coef_slp}{Coefficient for inter-annual change for the ith Landsat Band}

\item{coef_cos}{Coefficient for intra-annual change for the ith Landsat Band (1st harmonic)}

\item{coef_sin}{Coefficient for intra-annual change for the ith Landsat Band (1st harmonic)}

\item{coef_cos2}{Coefficient for intra-annual change for the ith Landsat Band (2nd harmonic)}

\item{coef_sin2}{Coefficient for intra-annual change for the ith Landsat Band (2nd harmonic)}

\item{coef_cos3}{Coefficient for intra-annual change for the ith Landsat Band (3rd harmonic)}

\item{coef_sin3}{Coefficient for intra-annual change for the ith Landsat Band (3rd harmonic)}
}
\value{
A floating point value for the calculated surface reflectance
}
\description{
For a given Julian date, and set of CCDC coefficients, this function returns
the calculated surface reflectance value using the full CCDC model described in
Zhu et al 2015. Called by other functions.
}
