% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccdc_current_seg.R
\name{date_to_jday}
\alias{date_to_jday}
\title{Simple function for converting a Date object to Julian date.}
\usage{
date_to_jday(date)
}
\arguments{
\item{date}{(Date) A R Date object to convert to Julian date.}
}
\value{
(integer) A integer representing the Julian date (i.e., days since '0000-01-01')
}
\description{
Simple function for converting a Date object to Julian date.
}
