% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdc_synthetic_img.R
\name{gen_ccdc_synthetic_img}
\alias{gen_ccdc_synthetic_img}
\title{Get the segment coefficients from the CCDC image at a specific Julian date (see Zhu & Woodcock 2014)}
\usage{
gen_ccdc_synthetic_img(ccdc_img, date, orig_crs = NULL)
}
\arguments{
\item{ccdc_img}{(SpatRaster, stars or data.frame) The CCDC image for which to extract coefficients as SpatRaster, stars or data.frame
with XY coordinates (i.e., terra::as.data.frame(img,xy=T)), data.frame is the fastest.}

\item{date}{(character) Date for which to extract CCDC coefficients (must be with in the date range for which the CCDC image was processed).}

\item{orig_crs}{(optional) In the case that a data.frame is provided this must be set to the original CRS of the image
in format recognized by 'Terra' package.}
}
\value{
(SpatRaster) A SpatRaster with all 7 Landsat bands predicted from the CCDC coefficients for the specified date.
}
\description{
Using the coefficients from the CCDC image a gap free Landsat Scene for any date (within the date range of the CCDC image)
can be generated using this function.
}
