% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccdc_current_seg.R
\name{gen_latest_ccdc_rast}
\alias{gen_latest_ccdc_rast}
\title{Get the most recent segment coefficients from the CCDC image (see Zhu & Woodcock 2014)}
\usage{
gen_latest_ccdc_rast(ccdc_img, max_seg, as_SpatRaster = T)
}
\arguments{
\item{ccdc_img}{(SpatRaster, stars) The CCDC image for which to extract coefficients.}

\item{max_seg}{(integer) The maximum number of segments present in the CCDC image.}

\item{as_SpatRaster}{(boolean) True or False, return the CCDC coefficient (and 'tStart')
as a SpatRaster image, otherwise a data frame. default TRUE.}
}
\value{
(SpatRaster) A SpatRaster with all CCDC coefficients (and tStart) associated with the most current segment.
If as_Spatrast is false, returns a data.frame.
}
\description{
For pixel locations where change has occurred (number of segments >1) this function returns
the model coefficients associated with the most recent segment, as well as the start
date of the segment as a Julian date. If there is only one segment (no change), returns coefficients
for the that segment.
}
