% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccdc_xy.R
\name{get_ccdc_ts}
\alias{get_ccdc_ts}
\title{Predict surface reflectance timeseries at a point (X,Y) using CCDC algorithm (Zhu & Woodcock 2014)}
\usage{
get_ccdc_ts(ccdc_img, x_coord, y_coord, epsg, band, n_seg = 10)
}
\arguments{
\item{ccdc_img}{(stars, SpatRaster) A multiband raster of the CCDC output from Google Earth Engine
with all coefficients for each segment present (i.e., see EE script)}

\item{x_coord}{(float) Value of the X coordinate of interest, must intersect the CCDC image}

\item{y_coord}{(float) Value of the Y coordinate of interest, must intersect the CCDC image}

\item{epsg}{(integer) EPSG code for CRS of the CCDC image}

\item{band}{(character) Band of interest, i.e., 'blue','green','red','nir','swir1','swir2' or 'therm'}

\item{n_seg}{(integer) Default 4. Number of segments to compute time series for, must be
less then or equal to maximum number of exported segments (i.e., see EE script)}
}
\value{
A 2-col matrix with the Julian date time series in the first and the
predicted CCDC time series in the second.
}
\description{
Provided a CCDC image output from the Google Earth Engine CCDC algorithm and
a X-Y coordinate this function returns the surface reflectance time series for
the intersected Landsat pixel.
}
