% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccdc_current_seg.R
\name{get_latest_seg_coefs}
\alias{get_latest_seg_coefs}
\title{Get the most recent segment coefficients from the CCDC image (see Zhu & Woodcock 2014)}
\usage{
get_latest_seg_coefs(ccdc_df, band, max_seg, xy = F)
}
\arguments{
\item{ccdc_df}{(Data.frame) The CCDC image as a data frame.}

\item{band}{(character) The name of the band for which to extract CCDC coefficients. One of 'blue', 'green', 'red'
'nir', 'swir1' or 'swir2'.}

\item{max_seg}{(integer) The maximum number of segments present in the CCDC image.}

\item{xy}{(boolean) True or False, return the CCDC coefficient data frame with xy coordinates, and 'tStart'}
}
\value{
(matrix) A matrix with the most current CCDC coefficients for the specified band at each pixel location,
and possibly xy coordinates and 'tStart'.
}
\description{
For pixel locations where change has occurred (number of segments >1) this function returns
the model coefficients associated with the most recent segment, as well as the start
date of the segment as a Julian date. If there is only one segment (no change), returns coefficients
for the that segment.
}
