% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccdc_current_seg.R
\name{jday_to_date}
\alias{jday_to_date}
\title{Simple function for converting a Julian date to Date.}
\usage{
jday_to_date(jday)
}
\arguments{
\item{jday}{(integer) Julian date to convert to Date.}
}
\value{
(Date) A Date corresponding to provided Julian date at UTC.
}
\description{
Simple function for converting a Julian date to Date.
}
