% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccdc_current_seg.R
\name{jdoy_to_date}
\alias{jdoy_to_date}
\title{Simple function for converting a Julian date to Date.}
\usage{
jdoy_to_date(jdoy)
}
\arguments{
\item{jdoy}{(integer) Julian date to convert to Date.}
}
\value{
(Date) A Date corresponding to provided Julian date at UTC.
}
\description{
Simple function for converting a Julian date to Date.
}
