% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdc_synthetic_img.R
\name{name_ccdc_bands}
\alias{name_ccdc_bands}
\title{Name missing CCDC image bands}
\usage{
name_ccdc_bands(ccdc_img, n_seg)
}
\arguments{
\item{ccdc_img}{(SpatRaster, stars) The CCDC image for which to name bands}

\item{n_seg}{(integer) The number of CCDC change segments exported by the Google Earth Engine script (must match ccdc_img)}
}
\value{
(SpatRaster) The same CCDC image used as input with missing band names correctly named.
}
\description{
Uses know naming convention to name the bands of a CCDC image exported from Google Earth Engine script in the case that they are missing.
Typically only needed for VRTs and CCDC exports of >5 segments.
}
