% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ccdc_ts.R
\name{sample_ccdc_by_catg}
\alias{sample_ccdc_by_catg}
\title{Uses SAGA-GIS and RSAGA to create N number of random samples with provided polygons, inherits polygon attributes. Points also get attributed
with the coefficient from a CCDC image for a given date. Points are returned as a data.frame. Useful function for
a large area of interest, i.e., as a alternative to 'plot_ccdc_ts_quartiles'.}
\usage{
sample_ccdc_by_catg(polygons_pth, ccdc_img_pth, pnts_count, sep_dist, env)
}
\arguments{
\item{polygons_pth}{(character) File path to polygon vector file with categories of interest.}

\item{ccdc_img_pth}{(character) File path to raster file (e.g. GeoTiff) CCDC image for given date (see 'gen_ccdc_at_jdoy')}

\item{pnts_count}{(integer) Total number of X-Y locations to sample throughout ALL polygons.}

\item{sep_dist}{(integer) Minimum distance in map units that two sample locations can be.}

\item{env}{(RSAGA environment) A RSAGA environment to use.}
}
\value{
(data.frame) A data.frame with the sample locations, attributed with the polygon data and CCDC grid coefficient values.
}
\description{
Uses SAGA-GIS and RSAGA to create N number of random samples with provided polygons, inherits polygon attributes. Points also get attributed
with the coefficient from a CCDC image for a given date. Points are returned as a data.frame. Useful function for
a large area of interest, i.e., as a alternative to 'plot_ccdc_ts_quartiles'.
}
