% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccdc_current_seg.R
\name{gen_ccdc_at_jdoy}
\alias{gen_ccdc_at_jdoy}
\title{Get the segment coefficients from the CCDC image at a specific Julian date (see Zhu & Woodcock 2014)}
\usage{
gen_ccdc_at_jdoy(ccdc_img, date)
}
\arguments{
\item{ccdc_img}{(SpatRaster, stars) The CCDC image for which to extract coefficients.}

\item{date}{(Date) Date for which to extract CCDC coefficients (must be later then March 1 1984, and earlier than present date).}
}
\value{
(SpatRaster) A SpatRaster with all CCDC coefficients (and tStart,tEnd) associated with the segment corresponding with the
specified Julian date. If the date is very recent, some data may be missing.
}
\description{
For pixel locations where change has occurred (number of segments >1) this function returns
the CCDC model coefficients associated with the segment that corresponds in time with the specified Julian date,
as well as the start date of the segment as a Julian date. If there is only one segment (no change), returns coefficients
for the that segment. !!Assumes that up to eight segments may exist (see Earth Engine script)!!
}
