% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ccdc_ts.R
\name{plot_ccdc_ts_bypoly}
\alias{plot_ccdc_ts_bypoly}
\title{Plot CCDC  (Zhu & Woodcock 2014) time series sample by polygon.}
\usage{
plot_ccdc_ts_bypoly(ccdc_img, zone_poly, catg_col, days = 730, band, N = 60)
}
\arguments{
\item{ccdc_img}{(SpatRaster,stars) The CCDC image for which to extract coefficients, for one segment only,
such as the output from 'gen_latest_ccdc_rast'.}

\item{zone_poly}{(SpatVector) Polygons for which to plot CCDC time series sample, by category.}

\item{catg_col}{(Character) The name of the column that defines the category of each polygon.}

\item{days}{The number of days for which to plot the CCDC time series, defaults to 730 (2 years)}

\item{band}{band (character) The name of the band for which to extract CCDC coefficients. One of 'blue', 'green', 'red'
'nir', 'swir1' or 'swir2'.}

\item{N}{The number of pixels to sample (and plot) for each category.}
}
\value{
(void) Plots the CCDC time series for a sample of pixels stratified by polygon category.
}
\description{
Provided a CCDC image and a polygon layer with some category field, this function plots a sample
of the pixels with each polygon grouped by category, indicated by the 'catg_col' parameter.
}
